config: Check if a theme repo is installed (#20)
Updates Repo::compare() to check if themes are already installed and
update the theme directory accordingly in repo config.

Adds Repo::installed_themes() which returns a list of all the existing
themes by their names. It looks for "themes" directory under a given
config directory path for installed themes. The symlinks are traversed
to check the destination to be a theme directory. The "current" theme
directory is ignored.