update: Add support for repos from local file
This adds support for repos with URL as a local file. This enables
creating custom theme repos locally without the need to download a repo
file.

The implementation adds `Update::update_repos()` which iterates through
the repos and fetches the themes. If the repo URL is a web URL, it
downloads it. If the URL is a local file address, it reads the files as
a repo.

Introduces `url` crate for URL parsing and construction.